\name{nlsy}
\alias{nlsy}
\docType{data}
\title{An example data set}
\usage{
data(nlsy)
}
\description{
Data set used in Liu & Zhang (2016). 

marijuana: binary; 1=used, 0=not used

gender: binary; 1=female, 0=male

smoke: binary; 1=smoke, 0=not smoke

residence: binary; 1=urban areas, 0=rural areas

peer: comprised score on peers life style; the higher score, the healthier the peers live.
}
